refacctor: Move classes with Metabox namespace to Helper
