+++ trunk/src/Helper/class-metabox-view.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Trasweb\Plugins\DisplayMetadata\Metabox;
+namespace Trasweb\Plugins\DisplayMetadata\Helper;
 
 use Trasweb\Plugins\DisplayMetadata\Model\Abstract_Model;
 use Trasweb\Plugins\DisplayMetadata\Plugin;
